+++ Editor/pipeline_material_processors.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 301799dcf10bc3442a288ea619fd6e58
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
